Fix build error from adding "android_builtin" import.
Calling `addImport` on a module with no `root_source_file` causes a build error. Zig exposes a module called "root", which is the `root_source_file`, and it seems that it is expected to exist when adding imports. This makes it necessary to check if the module has a `root_source_file` before trying to add the "android_builtin" import.